--- src/CasCap.Common.Testing/Logging/TestProvider.cs
@@ -1,66 +1,0 @@
-namespace Microsoft.Extensions.Logging;
-
-/// <summary><see cref="ILoggerProvider"/> that routes log output to xUnit's <see cref="ITestOutputHelper"/>.</summary>
-[ExcludeFromCodeCoverage]
-public class TestLogProvider(ITestOutputHelper testOutputHelper) : ILoggerProvider
-{
-    private readonly ConcurrentDictionary<string, TestLogger> _loggers = new(StringComparer.OrdinalIgnoreCase);
-
-    /// <inheritdoc/>
-    public ILogger CreateLogger(string categoryName)
-    {
-        return _loggers.GetOrAdd(categoryName, _ => new TestLogger(testOutputHelper));
-    }
-
-    /// <inheritdoc/>
-    public void Dispose()
-    {
-        Dispose(true);
-        GC.SuppressFinalize(this);
-    }
-
-    /// <summary>Releases resources used by this provider.</summary>
-    protected virtual void Dispose(bool disposing)
-    {
-        // Cleanup
-        _loggers.Clear();
-    }
-}
-
-[ExcludeFromCodeCoverage]
-class TestLogger(ITestOutputHelper output) : ILogger
-{
-    private readonly List<LogEntry> _entries = [];
-
-    public IReadOnlyCollection<LogEntry> GetLogs() => _entries.AsReadOnly();
-
-    public IDisposable? BeginScope<TState>(TState state) where TState : notnull => null!;
-
-    public bool IsEnabled(LogLevel logLevel) => true;
-
-    public void Log<TState>(LogLevel logLevel, EventId eventId, TState state, Exception? exception, Func<TState, Exception?, string> formatter)
-    {
-        var entry = new LogEntry(logLevel, formatter(state, exception));
-        _entries.Add(entry);
-        try
-        {
-            output.WriteLine(entry.ToString());
-        }
-        catch (InvalidOperationException)
-        {
-            //xUnit throws when no test is active (e.g. during async disposal)
-        }
-    }
-}
-
-[ExcludeFromCodeCoverage]
-class LogEntry(LogLevel level, string message)
-{
-    public DateTime Timestamp { get; } = DateTime.Now;
-
-    public LogLevel LogLevel { get; } = level;
-
-    public string Message { get; } = message;
-
-    public override string ToString() => $"{Timestamp:o}: {Message}";
-}
